move AzureSlider to common package
