+++ PluginsAndFeatures/azure-toolkit-for-intellij/src/com/microsoft/azure/toolkit/intellij/springcloud/component/SpringCloudAppConfigPanel.form
@@ -237,7 +237,7 @@
           <text value="Persistent storage:"/>
         </properties>
       </component>
-      <nested-form id="32dc5" form-file="com/microsoft/azure/toolkit/intellij/springcloud/properties/AzureSlider.form" binding="numInstance">
+      <nested-form id="32dc5" form-file="com/microsoft/azure/toolkit/intellij/common/component/AzureSlider.form" binding="numInstance">
         <constraints>
           <grid row="8" column="1" row-span="1" col-span="2" vsize-policy="3" hsize-policy="3" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
         </constraints>
